Tags incorporated into sign-up form and tag implementation changed
