+++ components/Tag/styles.module.scss
@@ -2,9 +2,11 @@
 
 .tag {
   background-color: $default-blue;
-  display: inline;
+  display: inline-block;
   color: $default-white;
   padding: 8px;
   border-radius: 3px;
   cursor: pointer;
+  margin-right: 10px;
+  margin-bottom: 5px;
 }
